+++ NEWS.md
@@ -1,3 +1,9 @@
+# kv 0.2.1
+
+* Add `kvi()`, `ki()`, `vi()` for iterating over various parts of keys, values
+  or indices.
+* All functions now utilize `.kvi()` in dot-kvi.R
+
 # kv 0.2.0 
 
 * Added consistent behavior for missing NAMES